Add Username Prompt on login, UserView
Moved AddUserName into a component, edited UserView and added tests.
Added a username prompt to the AllUserAccess and ClientAdmin wrappers
when user is logged in and doesn't have a username set. Also removed
requirement for First Name and Last Name.

Edits to eslint, user emails, and user methods to enact all these
changes.